move validation.py inside pyfeatures
